requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
openai>=1.68.2
python-dotenv
requests
